Registering TaskDirectoryStore in constructor instead of StoreConfig
